Rename checkauth file and fix module not found error
