+++ nodeapi/api/middleware/checkauth.js
